move: 결제 상품 타입과 카테고리의 패키지 이동에 따라 수정되는 파일
Related to: #34
