+++ bank-application/dev-bank-pay/src/main/java/org/bank/pay/core/payment/PaymentService.java
@@ -7,8 +7,8 @@ import org.bank.core.cash.PaymentProcessingException;
 import org.bank.core.dto.response.ResponseCodeV2;
 import org.bank.core.dto.response.ResponseDtoV2;
 import org.bank.core.payment.Product;
-import org.bank.pay.core.payment.product.VirtualCash;
-import org.bank.pay.core.payment.product.Category.CategoryType;
+import org.bank.pay.core.event.product.VirtualCash;
+import org.bank.pay.core.event.product.Category.CategoryType;
 import org.bank.pay.core.payment.client.PaymentClient;
 import org.bank.pay.core.producer.product.PurchasedEventPublisher;
 import org.bank.pay.global.exception.PaymentException;
