+++ bank-application/dev-bank-pay/src/main/java/org/bank/pay/core/payment/client/PaymentClient.java
@@ -3,7 +3,7 @@ package org.bank.pay.core.payment.client;
 import org.bank.core.auth.AuthClaims;
 import org.bank.core.dto.response.ResponseCodeV2;
 import org.bank.core.payment.Product;
-import org.bank.pay.core.payment.product.Category.CategoryType;
+import org.bank.pay.core.event.product.Category.CategoryType;
 import org.bank.pay.core.payment.PaymentCallbackProperties;
 import org.bank.pay.core.payment.PaymentDetail;
