+++ bank-application/dev-bank-pay/src/main/java/org/bank/pay/core/payment/naver/NaverPaymentRequest.java
@@ -4,7 +4,7 @@ import lombok.Data;
 import lombok.experimental.SuperBuilder;
 import org.bank.core.auth.AuthClaims;
 import org.bank.core.payment.Product;
-import org.bank.pay.core.payment.product.Category.CategoryType;
+import org.bank.pay.core.event.product.Category.CategoryType;
 
 import java.util.Collections;
 import java.util.List;
