+++ bank-domain/dev-bank-pay/src/main/java/org/bank/pay/core/event/product/VirtualCash.java
@@ -1,4 +1,4 @@
-package org.bank.pay.core.payment.product;
+package org.bank.pay.core.event.product;
 
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
